Add // Create all Seeders

+++ database/migrations/2025_06_07_211000_create_users_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->timestamp('emailVerifiedAt')->nullable();
             $table->string('password');
-            $table->integer('phoneNumber');
+            $table->bigInteger('phoneNumber');
             $table->rememberToken();
             $table->timestamps();
         });
